+++ Tests/InstanaAgentTests/Beacons/CoreBeacon/CoreBeaconFactoryTests.swift
@@ -106,7 +106,7 @@ class CoreBeaconFactoryTests: InstanaTestCase {
 
     func test_create_from_string() {
         // Given
-        let httpBody = "ab\t15702\nav\tunknown-version\nbid\tB5FAF31C-FE37-482E-97F2-20D49C506586\nbt\tBackendTracingID\nbi\tcom.apple.dt.xctest.tool\ncn\tNone\nct\tWifi\nd\t1578569955952\ndma\tApple\ndmo\tx86_64\nec\t1\nem\tA client or server connection was severed in the middle of an in-progress load.\net\tNetwork Connection Lost\nhm\tPOST\nhs\t200\nhu\thttps://www.example.com\nk\tKEY\np\tiOS\nosn\tiOS\nosv\t13.3\nagv\t1.0.4\nro\tfalse\nsid\t70BED140-D947-4EC7-ADE9-8F1F7C6955D0\nusi\t70BED140-D947-4EC7-ADE9-8F1F7C6955D0\nt\thttpRequest\nti\t1578569955952\nuf\tc\nul\ten\nvh\t1792\nvw\t828\nm_meta1\tV\nm_meta2\tL\nh_X_K\tV"
+        let httpBody = "ab\t15702\nav\tunknown-version\nbid\tB5FAF31C-FE37-482E-97F2-20D49C506586\nbt\tBackendTracingID\nbi\tcom.apple.dt.xctest.tool\ncn\tNone\nct\tWifi\nd\t1578569955952\ndma\tApple\ndmo\tx86_64\nec\t1\nem\tA client or server connection was severed in the middle of an in-progress load.\net\tNetwork Connection Lost\nhm\tPOST\nhs\t200\nhu\thttps://www.example.com\nk\tKEY\np\tiOS\nosn\tiOS\nosv\t13.3\nagv\t1.0.4\nro\t\(String(InstanaSystemUtils.isDeviceJailbroken))\nsid\t70BED140-D947-4EC7-ADE9-8F1F7C6955D0\nusi\t70BED140-D947-4EC7-ADE9-8F1F7C6955D0\nt\thttpRequest\nti\t1578569955952\nuf\tc\nul\ten\nvh\t1792\nvw\t828\nm_meta1\tV\nm_meta2\tL\nh_X_K\tV"
 
         // When
         let sut: CoreBeacon