+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.6.6
+- Fix client side crash symbolication for arm64e architecture
+- Fix unit test cases that failed in command line execution
+
 ## 1.6.5
 - Add crash to mobile feature list and send to Instana backend
 - Add user_session_id (usi) to beacons